Add char version of trace runner
There's a lot of code duplication here that could be omitted, but I was
already having a hard enough time fighting the C++ compiler with this
code, disregarding templating and other issues.

We also need a way to run the data structures individually against the
trace files for performance comparisons.  I don't know if that should go
in these files or in other ones.